[Bronze V] Title: 이상한 기호, Time: 124 ms, Memory: 14176 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -2,16 +2,16 @@ import java.io.*;
 import java.util.*;
 
 public class Main {
-	static int a,b;
+	static long a,b;
 
 	public static void main(String[] args) throws IOException {
 		// System.setIn(new FileInputStream("input.txt"));
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		
 		StringTokenizer st = new StringTokenizer(br.readLine());
-		a = Integer.parseInt(st.nextToken());
-		b = Integer.parseInt(st.nextToken());
+		a = Long.parseLong(st.nextToken());
+		b = Long.parseLong(st.nextToken());
 		
-		System.out.println((int)(Math.pow(a, 2) - Math.pow(b, 2)));
+		System.out.println((long)(Math.pow(a, 2) - Math.pow(b, 2)));
 	}
 }
